✨ Refactor: Daily 및 Weekly 보고서 타입 통합 및 관련 파일 수정
