+++ src/const/sourceIcons.tsx
@@ -1,4 +1,4 @@
-import { Evidence } from "@/types/dailyReport";
+import { Evidence } from "@/types/reportType";
 import { FileText, Github, Mail, Users } from "lucide-react";
 
 export const SOURCE_ICONS: Record<Evidence["source"], React.ReactNode> = {
