+++ src/utils/getSourceIcon.tsx
@@ -1,5 +1,5 @@
 import { SOURCE_ICONS } from "@/const/sourceIcons";
-import { Evidence } from "@/types/dailyReport";
+import { Evidence } from "@/types/reportType";
 
 export const getSourceIcon = (source: Evidence["source"]) => {
   return SOURCE_ICONS[source];
